Code cleanup Fix linux compilation

+++ code/unix/unix_main.c
@@ -854,7 +854,7 @@ static void* try_dlopen(const char* base, const char* gamedir, const char* fname
   return libHandle;
 }
 
-void *Sys_LoadDll( const char *name, intptr_t (**entryPoint)(int, ...),
+void *Sys_LoadDll( const char *name, intptr_t (**entryPoint)(intptr_t, ...),
                    intptr_t (*systemcalls)(intptr_t, ...) ) 
 {
   void *libHandle;
